Added cosplayitemnote interface
