+++ CosNet.API/Data/Repositories/ICosplayItemNoteRepository.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using CosNet.API.Entities;
+
+namespace CosNet.API.Data.Repositories
+{
+    public interface ICosplayNoteRepository
+    {
+        void AddCosplayNote(CosplayNote CosplayNote);
+        bool CosplayNoteExists(Guid CosplayNoteId);
+        void DeleteCosplayNote(Guid CosplayNoteId);
+        CosplayNote GetCosplayNote(Guid CosplayNoteId);
+        IEnumerable<CosplayNote> GetCosplayNotes(Guid cosplayId);
+        bool SaveChanges();
+        void UpdateCosplayNote(CosplayNote CosplayNote);
+    }
+}
